Implementation-PCA/Codes_NonAdaptive/requirements.txt
mtcnn==0.1.1
numpy==1.22.2
opencv_python==4.5.5.62
Pillow==9.1.1
